requirements.txt
streamlit==1.32.0
unstructured[all-docs]==0.12.5
faiss-cpu==1.8.0
sentence-transformers==2.5.1
langchain==0.1.13
langchain-community==0.0.29
groq==0.4.2
python-dotenv==1.0.1